move edit command
