+++ src/commands/remind.rs
@@ -13,6 +13,33 @@ use serde_json;
 
 use crate::Task;
 
+pub fn set_reminders<R: Reminder, S: AsRef<OsStr>>(taskwarrior_args: Vec<S>) -> io::Result<()> {
+    let output = String::from_utf8(
+        Command::new("task")
+            .args(taskwarrior_args)
+            .arg("+WAITING")
+            .arg("export")
+            .output()?
+            .stdout,
+    )
+    .unwrap();
+    let tasks = serde_json::from_str::<Vec<Task>>(&output).unwrap();
+    for task in tasks.into_iter() {
+        let wait = match task.wait {
+            None => continue,
+            Some(wait) => wait,
+        };
+        let has_reminder = task
+            .tags
+            .map_or(false, |tags| tags.contains(&"reminder".to_string()));
+        if !has_reminder {
+            R::add_reminder(&task.uuid, &task.description, wait.0)?;
+        }
+    }
+
+    Ok(())
+}
+
 pub trait Reminder {
     fn add_reminder(uuid: &str, title: &str, datetime: DateTime<Local>) -> io::Result<()>;
 }
@@ -48,39 +75,12 @@ end", date = datetime.format("%Y-%m-%d"), time = datetime.format("%H:%M"), title
     }
 }
 
-pub fn set_reminders<R: Reminder, S: AsRef<OsStr>>(taskwarrior_args: Vec<S>) -> io::Result<()> {
-    let output = String::from_utf8(
-        Command::new("task")
-            .args(taskwarrior_args)
-            .arg("+WAITING")
-            .arg("export")
-            .output()?
-            .stdout,
-    )
-    .unwrap();
-    let tasks = serde_json::from_str::<Vec<Task>>(&output).unwrap();
-    for task in tasks.into_iter() {
-        let wait = match task.wait {
-            None => continue,
-            Some(wait) => wait,
-        };
-        let has_reminder = task
-            .tags
-            .map_or(false, |tags| tags.contains(&"reminder".to_string()));
-        if !has_reminder {
-            R::add_reminder(&task.uuid, &task.description, wait.0)?;
-        }
-    }
-
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[test]
     fn test_set_reminders() -> io::Result<()> {
-        set_reminders::<MacReminder>(vec!["24"])
+        set_reminders::<MacReminder, _>(vec!["24"])
     }
 }
